refactor(links): move linkNodeFromSelection into helper
